requirements.txt
torch==1.8.1
gdown==3.13.0
PyYAML==5.4.1
transformers==4.6.1
underthesea==1.3.1
numpy==1.19.5
tqdm==4.61.0